+++ ybt_website/src/components/home-page/StorySection.jsx
