+++ ybt_website/src/pages/Home.jsx
@@ -1,5 +1,4 @@
-import SuccessStories from "../components/home-page/SuccessStories";
-import NavBar from "../components/Navbar";
+import StorySection from "../components/home-page/StorySection";
 import HeroBanner from "../components/home-page/HeroBanner";
 import TeamPreview from "../components/home-page/TeamPreview";
 
@@ -7,7 +6,7 @@ function Home() {
   return (
     <>
       <HeroBanner></HeroBanner>
-      <SuccessStories />
+      <StorySection />
       <TeamPreview />
     </>
   );
